feat: tweet 수정, 삭제 권한 체크

+++ controller/tweet.js
@@ -2,7 +2,9 @@ import * as tweetRepository from "../data/tweet.js";
 
 export async function getTweets(req, res) {
 	const { userName } = req.query;
-	const data = await (userName ? tweetRepository.getAllByUserName(userName) : tweetRepository.getAll());
+	const data = await (userName
+		? tweetRepository.getAllByUserName(userName)
+		: tweetRepository.getAll());
 
 	res.status(200).json(data);
 }
@@ -32,10 +34,20 @@ export async function createTweet(req, res) {
 export async function updateTweet(req, res) {
 	const { id } = req.params;
 	const { text } = req.body;
-	const tweet = await tweetRepository.update(id, text);
 
-	if (tweet) {
-		return res.status(200).json(tweet);
+	const tweet = await tweetRepository.getById(id);
+
+	if (!tweet)
+		return res.status(404).json({ message: "존재하지 않는 트윗입니다." });
+
+	if (parseInt(tweet.id) !== parseInt(req.userId)) {
+		return res.status(403).json({ message: "게시글 권한이 없습니다." });
+	}
+
+	const updatedTweet = await tweetRepository.update(id, text);
+
+	if (updatedTweet) {
+		return res.status(200).json(updatedTweet);
 	} else {
 		return res.status(404).json({ message: `Tweet ${id} not found` });
 	}
@@ -44,6 +56,16 @@ export async function updateTweet(req, res) {
 export async function deleteTweet(req, res) {
 	const { id } = req.params;
 	if (!id) return res.status(400).send("missing id");
+
+	const tweet = await tweetRepository.getById(id);
+
+	if (!tweet)
+		return res.status(404).json({ message: "존재하지 않는 트윗입니다." });
+
+	if (parseInt(tweet.id) !== parseInt(req.userId)) {
+		return res.status(403).json({ message: "게시글 권한이 없습니다." });
+	}
+
 	await tweetRepository.remove(id);
 
 	return res.sendStatus(204);
